This is a big one.
Added a positional and renderable interface to assist with determining which nodes are renderable and what to do with them.

Modified the scene renderer to loop through all nodes/childNodes and render only those that are renderable.
Fixed a bug in a Node3D where I didn't set a default scale.
Made Input accessible through a static call(this is probably temporary)
Fixed a bug in AstroEngine.pom where the manifest wasn't properly setting the main file.
Converted Projection to use Matrix4f instead of mine.
Added a MoveableCamera that utilizes its update function.
Modified the engine to no longer use Projects but to strictly check the SceneManager for the scene it's supposed to render. The nodes themselves now handle scene switching.